Finalize app for deployment: cleaned API routes, added download + loading UI, .gitignore, requirements.txt
