+++ .gitignore
@@ -1,3 +1,30 @@
-.venv/
+# Python
+__pycache__/
+*.pyc
+*.pyo
+*.pyd
+.Python
+
+# Virtual environments
 .venv/
 venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Node / React
+node_modules/
+build/
+dist/
+*.log
+
+# Environment variables
+*.env
+
+# OS / system files
+.DS_Store
+Thumbs.db
+
+# IDE / Editor
+.vscode/
+.idea/
