--- app/templates/instructions.html
@@ -1,7 +1,0 @@
-{% if instr_type == 'file_example' %}
-<h3>File Upload Instructions</h3>
-<p>Explain the CSV format, required column, etc.</p>
-{% elif instr_type == 'text_example' %}
-<h3>Text Input Instructions</h3>
-<p>Example input, how to delimit responses, etc.</p>
-{% endif %}
