--- app/templates/landing.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Home{% endblock %}
-
-{% block content %}
-<style>
-  
-    .full-height {
-        height: 100vh;
-    }
-</style>
-
-<div class="container-fluid full-height d-flex justify-content-center align-items-center">
-    <div class="text-center">
-        <h1 class="mb-4 ">
-            Automate Thematic Analysis
-        </h1>
-        <a href="{{ url_for('views.new_upload') }}" class="btn btn-lg btn-outline-dark ">
-            Start Analysis
-        </a>
-    </div>
-</div>
-{% endblock %}
