--- app/templates/new_upload.html
@@ -1,185 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Upload Data{% endblock %}
-
-
-{% block content %}
-
-<div class="container my-5 d-flex justify-content-center">
-    <div class="card shadow-md border rounded p-4" style="width: 28rem;">
-    <h3 class="mb-5" align="center">Choose an Upload Type</h3>
-
-    <!-- Nav tabs -->
-    <ul class="nav nav-tabs justify-content-center mb-4" role="tablist">
-        <li class="nav-item">
-            <a class="nav-link {% if active_tab == 'file' %}active{% endif %}" 
-                id="file-tab"
-                href="{{ url_for('views.new_upload', tab='file') }}" 
-                role="tab">File Upload</a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link {% if active_tab == 'text' %}active{% endif %}" 
-                id="text-tab"
-                href="{{ url_for('views.new_upload', tab='text') }}" 
-                role="tab">Text Input</a>
-        </li>
-    </ul>
-
-
-        <form method="POST" action="{{ url_for('views.new_upload', tab=active_tab) }}" enctype="multipart/form-data">
-
-        <div class="tab-content">
-            <!-- File Upload Tab -->
-            <div class="tab-pane fade {% if active_tab == 'file' %}show active{% endif %}"
-                 id="file-upload" role="tabpanel" 
-                 aria-labelledby="file-tab">
-                <p class="fs-1"><strong>Instructions:</strong> Upload your CSV file containing feedback, specify the column name that holds the
-                    responses, enter your desired themes with related seed words, and click Submit to begin the
-                    analysis.</p>
-
-                </p>
-
-                <div class="d-flex flex-column align-items-center">
-                    <label for="file" class="btn btn-dark btn-lg px-5 fw-bold border border-5">
-                         Choose File
-                        <input type="file" id="file" name="file" hidden>
-                    </label>
-                    <p id="file-name" class="text-muted small"></p>
-                </div>
-
-
-
-                <div class="form-group d-flex align-items-center">
-                    <label for="feedback_column" class="mb-0 mr-2">Name of column with feedback:</label>
-                    <input type="text" class="form-control" id="feedback_column" name="feedback_column" placeholder=""
-                         style="max-width: 120px; height: 30px;">
-                </div>
-
-                <p class="text-left">see <a href="{{ url_for('views.instructions', instr_type='file_example') }}">example</a></p>
-
-                <!-- <select class="form-control mb-3">
-                    <option>--Select Option--</option>
-                </select> -->
-            </div>
-
-            <!-- Text Input Tab -->
-            <div class="tab-pane fade {% if active_tab == 'text' %}show active{% endif %}" id="text-input" role="tabpanel" aria-labelledby="text-tab">
-                <p><strong>Instructions:</strong> Lorem ipsum dolor sit amet...</p>
-
-                <div class="form-group">
-                    <label for="text">Enter Text:</label>
-                    <textarea class="form-control" id="text" name="text" rows="6"
-                        placeholder="Solving New Problems, Finishing a Polished Project, Developed Perseverance, ..."></textarea>
-                </div>
-
-                <div class="form-group d-flex align-items-center">
-                    <label for="delimiter" class="mb-0 mr-2">Delimiter:</label>
-                    <input type="tisext" class="form-control mr-3" id="delimiter" name="delimiter" placeholder="" 
-                        style="max-width: 60px; height: 30px;">
-                    <p class="mb-0">see <a href="{{ url_for('views.instructions', instr_type='text_example') }}">example</a></p>
-                </div>
-
-                <select class="form-control mb-3">
-                    <option>--Select Option--</option>
-                </select>
-            </div>
-        </div>
-
-        <!-- <div id="theme-groups">
-            <div class="theme-group border p-3 mb-3">
-                <div class="form-group">
-                    <label>Theme 1</label>
-                    <input type="text" name="theme[0]" class="form-control"  >
-                </div>
-                <div class="form-group">
-                    <label>Seeds (comma-separated)</label>
-                    <input type="text" name="seeds[0]" class="form-control" placeholder="e.g. Exciting, Inspirational, Scary"
-                        >
-                </div>
-            </div>
-        </div>
-        
-        <button type="button" class="btn btn-light mb-3" id="add-theme">Add Another Theme</button>
--->
-
-        <div class="text-center">
-            <button type="submit" class="btn btn-primary">Submit</button>
-        </div>
-
-    </form>
-    </div>
-</div>
-
-        
-
-<!-- 
-        <div id="theme-groups">
-            <div class="theme-group border p-3 mb-3">
-                <div class="form-group">
-                    <label>Theme 1</label>
-                    <input type="text" name="theme[0]" class="form-control" placeholder="Enter theme name" required>
-                </div>
-                <div class="form-group">
-                    <label>Seeds (comma-separated)</label>
-                    <input type="text" name="seeds[0]" class="form-control" placeholder="e.g. Exciting, Inspirational, Scary" required>
-                </div>
-            </div>
-        </div>
-
-        <button type="button" class="btn btn-secondary mb-3" id="add-theme">Add Another Theme</button> -->
-
-
-       
-
-        
-
-
- <script>
-    // let themeCount = 1;
-    // const maxThemes = 4;
-
-    // document.getElementById('add-theme').addEventListener('click', function () {
-    //     if (themeCount >= maxThemes) return;
-
-    //     const container = document.getElementById('theme-groups');
-    //     const index = themeCount;
-
-    //     const group = document.createElement('div');
-    //     group.className = 'theme-group border p-3 mb-3';
-    //     group.innerHTML = `
-    //   <div class="form-group">
-    //     <label>Theme ${index + 1}</label>
-    //     <input type="text" name="theme[${index}]" class="form-control">
-    //   </div>
-    //   <div class="form-group">
-    //     <label>Seeds (comma-separated)</label>
-    //     <input type="text" name="seeds[${index}]" class="form-control" placeholder="e.g. Exciting, Inspirational, Scary">
-    //   </div>
-    // `;
-    //     container.appendChild(group);
-    //     themeCount++;
-    // });
-
-
-
-
-    const fileInput = document.getElementById('file');
-    const fileNameDisplay = document.getElementById('file-name');
-
-    fileInput.addEventListener('change', () => {
-        fileNameDisplay.textContent = fileInput.files[0]?.name || '';
-    });
-  
-
-
-</script> 
-
-
-
-
-
-
-{% endblock %}
-
-
-
