--- app/templates/processing.html
@@ -1,22 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-Processing...
-{% endblock %}
-
-{% block content %}
-<div class="d-flex flex-column justify-content-center align-items-center" style="height: 80vh;">
-    <div class="spinner-border text-primary" role="status" style="width: 4rem; height: 4rem;">
-        <span class="sr-only">Loading...</span>
-    </div>
-    <h3 class="mt-4">Processing your submission...</h3>
-</div>
-
-<script>
-    // Run clustering in background, then redirect to results
-    fetch("/run-clustering/{{ public_id }}")
-        .then(() => {
-            window.location.href = "/results/{{ public_id }}";
-        });
-</script>
-{% endblock %}
