+++ backend/main.py
@@ -1,0 +1,7 @@
+from app import create_app
+
+
+app = create_app()
+
+if __name__ == '__main__':
+    app.run(port=5001)
