+++ backend/requirements.txt
@@ -1,0 +1,68 @@
+annotated-types==0.7.0
+anyio==4.9.0
+blinker==1.9.0
+certifi==2025.6.15
+charset-normalizer==3.4.2
+click==8.2.1
+contourpy==1.3.2
+cycler==0.12.1
+distro==1.9.0
+exceptiongroup==1.3.0
+filelock==3.18.0
+Flask==3.1.1
+Flask-SQLAlchemy==3.1.1
+Flask-Migrate
+fonttools==4.58.4
+fsspec==2025.5.1
+greenlet==3.2.3
+h11==0.16.0
+hf-xet==1.1.5
+httpcore==1.0.9
+httpx==0.28.1
+huggingface-hub==0.33.2
+idna==3.10
+itsdangerous==2.2.0
+Jinja2==3.1.6
+jiter==0.10.0
+joblib==1.5.1
+kiwisolver==1.4.8
+MarkupSafe==3.0.2
+matplotlib==3.10.3
+matplotlib-venn==1.1.2
+mpmath==1.3.0
+networkx==3.4.2
+numpy==1.23.5
+openai==1.93.0
+packaging==25.0
+pandas==2.3.0
+pillow==11.3.0
+pydantic==2.11.7
+pydantic_core==2.33.2
+pyparsing==3.2.3
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.1
+pytz==2025.2
+PyYAML==6.0.2
+regex==2024.11.6
+requests==2.32.4
+safetensors==0.5.3
+scikit-learn==1.7.0
+scipy==1.15.3
+seaborn==0.13.2
+sentence-transformers==5.0.0
+six==1.17.0
+sniffio==1.3.1
+SQLAlchemy==2.0.41
+sympy==1.14.0
+threadpoolctl==3.6.0
+tokenizers==0.21.2
+torch==2.2.2
+tqdm==4.67.1
+transformers==4.53.0
+typing-inspection==0.4.1
+typing_extensions==4.14.0
+tzdata==2025.2
+urllib3==2.5.0
+Werkzeug==3.1.3
+wordcloud==1.9.4
+WTForms==3.2.1
