+++ frontend/src/App.js
@@ -5,13 +5,14 @@ import ReviewThemes from './pages/ReviewThemes';
 import ResultsPage from './pages/ResultsPage';
 import ClusterExistingCodes from './pages/ClusterCodes';
 import DeductiveApproach from './pages/DeductiveApproach';
-
+import HomePage from './pages/HomePage';
 function App() {
   return (
     <>
       
       <TabNavigation /> {/* shows the nav bar with tab links */}
       <Routes>
+        <Route path="/" element={<HomePage />} /> 
         <Route path="/upload" element={<UploadPage />} />
         <Route path="/submission/:public_id/themes" element={<ReviewThemes />} />
         <Route path="/results/:public_id" element={<ResultsPage />} />
